+++ tests/metrics/test_primitives.py
@@ -4,7 +4,7 @@ import torch
 from torch.autograd import Variable
 
 import torchbearer
-from torchbearer.metrics import Loss, Epoch, CategoricalAccuracy, TopKCategoricalAccuracy
+from torchbearer.metrics import Loss, Epoch, CategoricalAccuracy, TopKCategoricalAccuracy, BinaryAccuracy
 
 
 class TestLoss(unittest.TestCase):
@@ -40,6 +40,44 @@ class TestEpoch(unittest.TestCase):
     def test_process_final(self):
         result = self._metric.process_final(self._state)
         self.assertEqual(101, result)
+
+
+class TestBinaryAccuracy(unittest.TestCase):
+    def setUp(self):
+        self._state = {
+            torchbearer.Y_TRUE: torch.LongTensor([
+                [1, 0, 0],
+                [0, 1, 0],
+                [0, 0, 1],
+                [0, 0, 1],
+                [0, 1, 0]
+            ]),
+            torchbearer.Y_PRED: torch.FloatTensor([
+                [0.9, 0.1, 0.1],  # Correct
+                [0.1, 0.9, 0.1],  # Correct
+                [0.1, 0.1, 0.9],  # Correct
+                [0.9, 0.1, 0.1],  # Incorrect
+                [0.9, 0.1, 0.1]  # Incorrect
+            ])
+        }
+        self._targets = [1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1]
+        self._metric = BinaryAccuracy().root  # Get root node of Tree for testing
+
+    def test_train_process(self):
+        self._metric.train()
+        result = self._metric.process(self._state)
+        for i in range(0, len(self._targets)):
+            self.assertEqual(result[i], self._targets[i],
+                             msg='returned: ' + str(result[i]) + ' expected: ' + str(self._targets[i])
+                                 + ' in: ' + str(result))
+
+    def test_validate_process(self):
+        self._metric.eval()
+        result = self._metric.process(self._state)
+        for i in range(0, len(self._targets)):
+            self.assertEqual(result[i], self._targets[i],
+                             msg='returned: ' + str(result[i]) + ' expected: ' + str(self._targets[i])
+                                 + ' in: ' + str(result))
 
 
 class TestCategoricalAccuracy(unittest.TestCase):