+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added the ability to pass custom arguments to the tqdm callback
 - Added an ignore_index flag to the categorical accuracy metric, similar to nn.CrossEntropyLoss. Usage: ``metrics=[CategoricalAccuracyFactory(ignore_index=0)]``
-- Added TopKCategoricalAccuracyMetric
+- Added TopKCategoricalAccuracy metric
+- Added BinaryAccuracy metric 
 ### Changed
 ### Deprecated
 ### Removed